+++ R/simulate.R
@@ -381,6 +381,14 @@ simulate_atomic = function(fit,
   lapply(args, checkmate::assert_numeric, any.missing = FALSE)
   lapply(args, function(x) stopifnot(length(x) == 1 | length(x) == nrow(newdata)))
 
+  if (.type == "predict" && .arma) {
+    warn_high_order_arma(
+      get_arma_order(fit$.internal$rhs_table, "ar"),
+      get_arma_order(fit$.internal$rhs_table, "ma"),
+      "simulate"
+    )
+  }
+
   # Remove response column if present - it is to be simulated
   if (fit$pars$y %in% colnames(newdata))
     newdata = dplyr::select(newdata, -dplyr::all_of(fit$pars$y))